requirements.txt
asgiref==3.6.0
Django==4.2.6
django-environ==0.11.2
django-filter==22.1
#django-quill-editor==0.1.40
django-simple-history==3.2.0
gunicorn==20.1.0
python-dotenv==0.21.1
sqlparse==0.4.2
tzdata==2023.3
whitenoise==5.3.0
